+++ gateway-js/CHANGELOG.md
@@ -1,5 +1,14 @@
 # CHANGELOG for `@apollo/gateway`
 
+## 2.8.0-connectors.3
+
+### Patch Changes
+
+- Updated dependencies []:
+  - @apollo/federation-internals@2.8.0-connectors.3
+  - @apollo/composition@2.8.0-connectors.3
+  - @apollo/query-planner@2.8.0-connectors.3
+
 ## 2.8.0-connectors.2
 
 ### Patch Changes