+++ query-planner-js/CHANGELOG.md
@@ -1,5 +1,13 @@
 # CHANGELOG for `@apollo/query-planner`
 
+## 2.8.0-connectors.3
+
+### Patch Changes
+
+- Updated dependencies []:
+  - @apollo/federation-internals@2.8.0-connectors.3
+  - @apollo/query-graphs@2.8.0-connectors.3
+
 ## 2.8.0-connectors.2
 
 ### Patch Changes